requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.7.0

# OCR

pytesseract==0.3.10
Pillow==10.2.0

# For parsing and regex

regex

# (Optional) If you want to allow file uploads in Postman

python-multipart==0.0.9

# (Optional) For testing API calls

requests==2.31.0
